+++ docs-src/components/popover/Popover.svelte
@@ -20,6 +20,19 @@
 </Popover>
 
 
+<h3>No tip</h3>
+<p>Styling is different than the normal popover, because the use-case for no-tip popover
+	is more similar to a dropdown rather than a tooltip or a popover,
+	so it makes sense that it also looks for the role.</p>
+
+<Button on:click="{popover5.open}">Open popover</Button>
+<Popover hideTip bind:this="{popover5}">
+	<h2>Context information</h2>
+	<p>Some text</p>
+	<Button on:click="{popover5.close}">Click me</Button>
+</Popover>
+
+
 <h3>Custom offset</h3>
 <Button on:click="{popover2.open}">Open popover</Button>
 <Popover bind:this="{popover2}" offset="-20">Smaller offset</Popover>
@@ -48,7 +61,7 @@ import { Popover, Button } from '../../../src';
 import { API } from '../../api-table';
 import { CodeExample } from '../../code-example';
 
-let popover1, popover2, popover3, popover4;
+let popover1, popover2, popover3, popover4, popover5;
 let content = '<h2>Context information</h2><p>Some text</p>';
 
 function updateContent () {
@@ -58,6 +71,7 @@ function updateContent () {
 const apiProps = [
 	{ name: 'class', type: 'string', description: 'Additional css class name to be added to the component.' },
 	{ name: 'offset', type: 'number', default: '2', description: 'Customize popover offset. Use negative number for smaller offset or positive for bigger' },
+	{ name: 'hideTip', description: 'Display just the container, without the tip (small triangle pointing at the target).' },
 	{ name: 'position', type: ['top', 'bottom'], default: 'bottom', description: 'Prefer the position of the popover to be above (top) or below (bottom) the target element.' },
 	{ name: 'bind:element', type: 'element', description: 'Exposes the HTML element of the component.' },
 	{ name: 'bind:contentElement', type: 'element', description: 'Exposes the HTML element of the content div.' },